pyproject.toml
[project]
name = "agentic-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=0.5.0",
    "cohere>=5.20.7",
    "docling>=2.81.0",
    "faiss-cpu>=1.13.2",
    "google-cloud-aiplatform>=1.50.0",
    "groq>=0.9.0",
    "langchain>=0.2.0",
    "langchain-community>=0.2.0",
    "langchain-core>=0.2.0",
    "langchain-google-vertexai>=3.2.2",
    "langchain-groq>=1.1.2",
    "langgraph>=0.1.0",
    "openai>=1.30.0",
    "pymupdf>=1.24.0",
    "pymupdf4llm>=0.0.17",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=2.7.0",
    "tavily>=1.1.0",
    "torch>=2.0.0",
    "transformers>=4.40.0",
]
